+++ hcore/harray.h
@@ -80,7 +80,7 @@ HArray < tType >::HArray ( int a_iSize ) : f_iSize ( 0 ), f_ptArray ( NULL )
 	f_iSize = a_iSize;
 	if ( a_iSize )
 		{
-		f_ptArray = new ( std::nothrow ) tType [ static_cast < unsigned int > ( f_iSize ) ];
+		f_ptArray = new ( std::nothrow ) tType [ f_iSize ];
 		if ( ! f_ptArray )
 			M_THROW ( g_ppcErrMsgHArray [ E_HARRAY_NOMEM ], a_iSize );
 		}
@@ -99,7 +99,7 @@ HArray < tType >::HArray ( const int & a_iSize, tType const & a_tFillWith )
 	f_iSize = a_iSize;
 	if ( a_iSize )
 		{
-		f_ptArray = new ( std::nothrow ) tType [ static_cast < unsigned int > ( f_iSize ) ];
+		f_ptArray = new ( std::nothrow ) tType [ f_iSize ];
 		if ( ! f_ptArray )
 			M_THROW ( g_ppcErrMsgHArray [ E_HARRAY_NOMEM ], a_iSize );
 		for ( l_iCtr = 0; l_iCtr < f_iSize; l_iCtr ++ )
@@ -149,7 +149,7 @@ HArray < tType > & HArray < tType >::operator = ( const HArray & a_roArray )
 			}
 		if ( f_iSize && ! f_ptArray )
 			{
-			f_ptArray = new ( std::nothrow ) tType [ static_cast < unsigned int > ( f_iSize ) ];
+			f_ptArray = new ( std::nothrow ) tType [ f_iSize ];
 			if ( ! f_ptArray )
 				M_THROW ( g_ppcErrMsgHArray [ E_HARRAY_NOMEM ], f_iSize );
 			}
